finish fetchEcosystemsFromStart tests
