+++ defi/src/cli/backfillDexVolumes/fetchAllEcosystemsFromStart/fetchAllEcosystemsFromStart.test.ts
@@ -1,24 +1,16 @@
 import fetchAllEcosystemsFromStart from "./";
 
 import { fetchEcosystemsFromStart } from "../";
-import { getDexVolumeRecord } from "../../../dexVolumes/dexVolumeRecords";
 
 import {
+  adapterFetch,
   fetchEcosystemsFromStartResult,
   fetchAllEcosystemsFromStartResult,
+  traderJoeVolumeAdapter,
 } from "../fixtures";
-
-import { ChainBlocks } from "../../../dexVolumes/dexVolume.types";
 
 afterEach(() => {
   jest.clearAllMocks();
-});
-
-const fetch = async (_timestamp: number, _chainBlocks: ChainBlocks) => ({
-  block: 1,
-  dailyVolume: "0",
-  totalVolume: "0",
-  timestamp: 1,
 });
 
 jest.mock("../", () => {
@@ -42,35 +34,12 @@ test("fetchEcosystemsFromStart", async () => {
     start,
     ecosystem,
     end,
-    fetch,
+    fetch: adapterFetch,
   });
   expect(mockedFetchEcosystemsFromStart).toEqual(
     fetchEcosystemsFromStartResult
   );
   expect(fetchEcosystemsFromStart).toHaveBeenCalled();
-});
-
-const traderjoeId = 1;
-const traderjoe = { module: "traderjoe" };
-const uniswapId = 1;
-const uniswap = { module: "uniswap" };
-
-jest.mock("../../../dexVolumes/dexVolumeRecords", () => {
-  const originalModule = jest.requireActual(
-    "../../../dexVolumes/dexVolumeRecords"
-  );
-
-  return {
-    __esModule: true,
-    ...originalModule,
-    getDexVolumeRecord: jest.fn((id) => (id === 1 ? traderjoe : uniswap)),
-  };
-});
-
-test("getDexVolumeRecord", async () => {
-  const mockedGetDexVolumeRecord = getDexVolumeRecord(traderjoeId);
-  expect(mockedGetDexVolumeRecord).toEqual(traderjoe);
-  expect(getDexVolumeRecord).toHaveBeenCalled();
 });
 
 jest.mock("../../../../DefiLlama-Adapters/dexVolumes", () => {
@@ -84,8 +53,10 @@ describe("fetchAllEcosystemsFromStart", () => {
   const end = 1642370400; // 16/1/22 22:00
   describe("Adapter with volume", () => {
     it("return all ecosystems with their respective volumes", async () => {
-      const result = await fetchAllEcosystemsFromStart(traderjoeId, end);
-      expect(getDexVolumeRecord).toHaveBeenCalled();
+      const result = await fetchAllEcosystemsFromStart(
+        traderJoeVolumeAdapter.volume,
+        end
+      );
       expect(fetchEcosystemsFromStart).toHaveBeenCalledTimes(2);
 
       expect(result).toEqual(fetchAllEcosystemsFromStartResult);
